Remove unnecessary Zone.Identifier files and update Recebimento model properties
This commit removes the unnecessary Zone.Identifier files from the python directory. These files were not needed and were causing clutter in the repository.

Additionally, the commit updates the properties of the Recebimento model in the src/app/models/recebimento.model.ts file. The dtPgto property has been renamed to data, and the recibo property has been renamed to id_recibo. These changes improve the clarity and consistency of the model's properties.