refactor: Update RegistraDebitoUniforme component
The code changes in the RegistraDebitoUniforme component include updating the HTML template and TypeScript file. The changes involve replacing the mat-autocomplete references with more descriptive names, 'auto' with 'dbv' and 'auto' with 'unif'. This improves the clarity and readability of the code.

This commit message follows the convention of starting with a verb in the imperative form ('Update', 'Fix', etc.) to describe the purpose of the changes.